requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.36
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.17.0
